Renames edit to page-edit
