+++ client/pages/index.js
@@ -1,6 +1,6 @@
 import Layout from './layout'
 import Home from './home'
-import Edit from './edit'
+import PageEdit from './page-edit'
 import Project from './project'
 import Page from './page'
 import Facts from './facts'
@@ -9,7 +9,7 @@ import FactsEdit from './facts-edit'
 export {
   Layout,
   Home,
-  Edit,
+  PageEdit,
   Project,
   Page,
   Facts,
