+++ client/routes.js
@@ -44,7 +44,7 @@ FlowRouter.route('/page/:pageId', {
       })
     } else {
       mount(pages.Layout, {
-        content: React.createElement(pages.Edit, {pageId: params.pageId, field: queryParams.field})
+        content: React.createElement(pages.PageEdit, {pageId: params.pageId, field: queryParams.field})
       })
     }
   }
